requirements.txt
APScheduler==3.11.0
asgiref==3.10.0
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.4
css-html-js-minify==2.5.5
cssselect==1.3.0
cssutils==2.11.1
Django==4.2.11
django-apscheduler==0.7.0
idna==3.11
Jinja2==3.1.6
lxml==6.0.2
MarkupSafe==3.0.3
more-itertools==10.8.0
numpy==2.3.3
pandas==2.3.3
pillow==11.3.0
premailer==3.10.0
PyMySQL==1.1.2
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
reportlab==4.4.4
reportlib==3.4.0
requests==2.32.5
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
PyMuPDF==1.24.10
pytesseract==0.3.13

